[UPD] Analytic changes into xe_pacific module
