--- xe_analytic/models/__init__.py
@@ -1,5 +1,0 @@
-from . import analytic_distribution_model
-from . import sale_order_line
-from . import purchase_order_line
-from . import account_move
-from . import account_move_line
